Add files via upload
This Notebook is a work in progress.
The function text_to_df() works in that when called on an already downloaded file from the CAISO website it returns the correct output. However, when either the urllib or requests package is used to automate successive downloads of this daily report (with the goal of making an annual series) , then it misreads the file oddly (error message includes "sees on Line 6, 28 columns instead of the expected 6". Why on line 6 rather than line 3 after the skiprows?  Why was 6 expected when the number of columns should have been expected to have been 8. and so on. In fact sometimes the file looks like complete garbage.
This is very puzzling. Moreover, the older urllib package at least downloads a file while the supposedly better and newer requests approach does not even do that.
One perhaps wild idea might be to use Beautiful soup and .findall() to get the urls of about five years of daily reports.
Wondering rather painfully whether to cut my losses and just download a few months manually!
Constructive suggestions much appreciated as how to proceed.