Add files via upload
This Notebook develops three function, namely 1. get_price_series(files), 2. get_hours_histogram(files), and 3. get_OLS(files)
The files are a list of CSV downloads from the CAISO website, which cover 2017 & 2018 in monthly data. The prices are hourly from the so-called "Day-Ahead-Market" (DAM), since most transactions are conducted on this market, and the Spot Market (RTM) is much less used.
#1 returns an hourly price series for the PG&E main node.
#2 returns a histogram for which hour of the day had the lowest price.
#3 returns an OLS regression for the subject time period embodied by the files.